Move ImageAlignment into a module
