+++ metaspace/webapp/src/modules/GroupProfile/EditGroupProfile.vue
@@ -68,7 +68,7 @@
   } from '../../api/group';
   import gql from 'graphql-tag';
   import EditGroupForm from './EditGroupForm.vue';
-  import MembersList from '../../components/MembersList/MembersList.vue';
+  import MembersList from '../../components/MembersList.vue';
   import { UserRole } from '../../api/user';
   import { encodeParams } from '../Filters';
   import ConfirmAsync from '../../components/ConfirmAsync';
